+++ apps/image-generation/src/index.ts
@@ -1,7 +1,7 @@
 import { createApp } from './controller'
 import { adminApp } from './controller/devPage'
-import { imageMetadataRepoCloudflareD1WithKV } from './infra/repos/image-metadata-cf-d1-kv'
-import { imageStoreRepoCloudflareR2 } from './infra/repos/image-store-cf-r2'
+import { imageMetadataRepoCloudflareD1WithKV } from './infra/repos/imageMetadataRepoCloudflareD1WithKV'
+import { imageStoreRepoCloudflareR2 } from './infra/repos/imageStoreRepoCloudflareR2'
 import { createOpenAIChatLLMJson } from './infra/services/llm'
 import { randomWordApi } from './infra/services/random-words'
 import { createHfImageGenerator } from './infra/services/text-to-image'
