+++ apps/image-generation/src/infra/repos/imageMetadataRepoCloudflareD1WithKV.ts
