+++ apps/image-generation/src/infra/repos/imageStoreRepoCloudflareR2.ts
